These are the 2 gravity simulations I have made.
The GravitySim.jar is a 2d simulation and the GravitySim3D.jar is a 3d simulation.

They are both runnable jar files. They work by clicking on the screen to place a planet. The options menu lets you edit the mass, velocity, and z position of the planet. Lots of planets will start to lag since the programs calculates the gravitational influence of each planet on each other planet.